# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Recherche d'emploi",
        "description": "Documentation de l'API Laravel pour la plateforme de recherche d'emploi",
        "contact": {
            "email": "hariznabil663@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur principal"
        }
    ],
    "paths": {
        "/api/show-fans": {
            "get": {
                "tags": [
                    "CRUD _Users"
                ],
                "summary": "Afficher tous les utilisateurs avec leurs rôles",
                "description": "Retourne la liste des utilisateurs avec leurs rôles associés (accessible uniquement aux administrateurs).",
                "operationId": "079729b90e2da907adac9705f0ec84fe",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "first_name": {
                                                "type": "string",
                                                "example": "Nabil"
                                            },
                                            "last_name": {
                                                "type": "string",
                                                "example": "Hariz"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "nabil@test.ma"
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "0612345678"
                                            },
                                            "roles": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "example": "admin"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé — Token ou rôle invalide"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/add-fan": {
            "post": {
                "tags": [
                    "CRUD _Users"
                ],
                "summary": "Créer un nouveau fan",
                "operationId": "29f33cf0fc3871dfa7a81484bf29980b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Nabil"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Hariz"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "nabil@test.ma"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0612345678"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Fan créé avec succès"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/fan/{id}": {
            "put": {
                "tags": [
                    "CRUD _Users"
                ],
                "summary": "Mettre à jour un fan",
                "operationId": "3d25401571c9618757b8a5d7a206b808",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du fan à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Nabil"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Hariz"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "nabil@test.ma"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0612345678"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "newpassword123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Fan mis à jour avec succès"
                    },
                    "404": {
                        "description": "Fan non trouvé"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "CRUD _Users"
                ],
                "summary": "Supprimer un fan",
                "operationId": "d3d3170de8b53b858c8fc866ce77336a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du fan à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fan supprimé avec succès"
                    },
                    "404": {
                        "description": "Fan non trouvé"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/fan-details/{id}": {
            "get": {
                "tags": [
                    "CRUD _Users"
                ],
                "summary": "Afficher le détail d'un fan",
                "operationId": "c49b9376ac378c4d3f3ce4311c9a3072",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du fan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails récupérés avec succès"
                    },
                    "404": {
                        "description": "Fan non trouvé"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}